### 2024-04-24 #6674 @editorialbot
[PRE REVIEW]: mde: An R package to ease missing data exploration and handling
**Submitting author:** @Nelson-Gon (<a href="http://orcid.org/0000-0003-3360-3396">Nelson Gonzabato</a>)
**Repository:** https://github.com/Nelson-Gon/mde
**Branch with paper.md** (empty if default branch): joss_submission
**Version:** v0.3.2
