### 2024-04-25 #6674 @crvernon
In the meantime, could you fix the following:

> ⚠️ [An error happened when generating the pdf](https://github.com/openjournals/joss-papers/actions/runs/8816732430?check_suite_focus=true). Problem with ORCID (https://orcid.org/0000-0003-3360-3396) for Nelson Gonzabato. ORCID looks to be the wrong length.

and the invalid DOI here:

> ```
> Reference check summary (note 'MISSING' DOIs are suggestions that need verification):
>
> OK DOIs
>
> - 10.18637/jss.v105.i07 is OK
> - 10.21105/joss.00355 is OK
> - 10.18637/jss.v045.i03 is OK
>
> MISSING DOIs
>
> - No DOI given, and none found for title: R: A Language and Environment for Statistical Comp...
> - No DOI given, and none found for title: dplyr: A Grammar of Data Manipulation
> - No DOI given, and none found for title: tidyr: Tidy Messy Data
>
> INVALID DOIs
>
> - https://doi.org/10.1016/j.jsp.2009.10.001 is INVALID because of 'https://doi.org/' prefix
> ```
