### 2024-04-28 #6674 @EugeneHao
Hi Taher,

Thank you for your email! I am sorry that I will not be available in the
next few weeks to review this paper.

Best,

Hao

On Fri, Apr 26, 2024 at 8:41 PM Taher Chegini ***@***.***>
wrote:

> Hi @EugeneHao <https://github.com/EugeneHao> and @roualdes
> <https://github.com/roualdes>! Would you like to review this submission
> to the Journal for Open Source Software? Our reviews are checklist-driven
> and openly conducted on GitHub over a timeline of 4–6 weeks. Because the
> process is much more iterative and interactive than a traditional paper
> review, we would ask you to start within the next 2 weeks. Here are
> reviewer guidelines for reference:
> joss.readthedocs.io/en/latest/reviewer_guidelines.html
>
> Thanks for your consideration.
>
