### 2024-04-29 #6697 @editorialbot
[REVIEW]: mde: An R package to ease missing dat exploration and handling
**Submitting author:** @Nelson-Gon (<a href="http://orcid.org/0000-0003-3360-3396">Nelson Gonzabato</a>)
**Repository:** https://github.com/Nelson-Gon/mde
**Branch with paper.md** (empty if default branch): joss_submission
**Version:** v0.3.2
