### 2024-05-28 #6697 @roualdes
A good idea here, to provide some helper functions for working with and understanding the scope of missing data within data frames.  I'm just a little hung up on the **Substantial scholarly effort** checkbox.

For instance, [cloc](https://github.com/AlDanial/cloc/) counts roughly 650 lines of code (excluding tests, blank lines, comments, and doc strings).

On the positive side of things, for sure some functions look like re-usable patterns and combinations of dplyr tooling.

On the negative side of things, however, some of the functions are wrappers around pretty standard patterns in R, e.g. `sum(is.na(x))` for counting NAs, and other functions are just catching errors that dplyr already catches.

Somewhere in between positive and negative, many of these functions could be cleaned up.

* `drop_all_na.data.frame` drops columns or rows of the user supplied data frame depending on second argument.  The doc however suggests only columns will be dropped.  From a user's perspective, this seems to break [POLA](https://en.wikipedia.org/wiki/Principle_of_least_astonishment).
* `recode_as_na_for.data.frame` is pretty computationally inefficient, performing way too much work.  Most of this function could be reduced to just a couple of lines of code abusing the fact that R can pass operators as function arguments, see `?match.fun`.
* Comments like "This is currently not the best way, a prototype that JustWorks^^TM" aren't very encouraging.
* The code is oddly formatted, with the body of functions sometimes being indented, sometimes not.  And an inconsistent use of assignment operators, e.g. <-, =, ->.

I'll stop here for now and ask for guidance on how best to proceed.  Please advise @cheginit.  Thanks.
